requirements.txt
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
ipython==8.24.0
jedi==0.19.1
jellyfish==1.0.3
joblib==1.4.2
lxml==5.2.2
matplotlib-inline==0.1.7
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
parso==0.8.4
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycodestyle==2.11.1
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
recordlinkage==0.16
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
wcwidth==0.2.13
